Add PWA runtime caching with offline fallback
