+++ public/fallback.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8" />
+  <title>Auricle Offline</title>
+  <style>
+    body { font-family: sans-serif; padding: 2rem; text-align: center; }
+  </style>
+</head>
+<body>
+  <h1>You’re offline</h1>
+  <p>Please reconnect to continue.</p>
+</body>
+</html>
